98 Fibonacci numbers, starting with 1 and 2, followed by a new line.

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,17 +6,17 @@
 */
 int main(void)
 {
-unsigned long a = 1, b = 2, c;
+unsigned long long a = 1, b = 2, c;
 int count = 0;
 
 while (count < 98)
 {
 if (count == 0)
-printf("%lu, %lu, ", a, b);
+printf("%llu, %llu, ", a, b);
 else if (count != 97)  /* Print the comma and space */
-printf("%lu, ", c);
+printf("%llu, ", c);
 else                  /* Print the last number with a newline */
-printf("%lu\n", c);
+printf("%llu\n", c);
 
 /* Calculate the next Fibonacci number */
 c = a + b;
